투사체가 충돌시 ATriggerVolume을 무시하고 투사체의 소유자(Weapon)과 소유자의 소유자(Pioneer)는 무시하고 투사체끼리도 무시하도록 if(OtherActor->IsA(AProjectile::StaticClass()) 검사를 함. C++ 폴더생성 및 코드 정리.투사체와 Enemy가 충돌시 Enemy가 바로 사라지지 않고 HP가 투사체의 소유자(Weapon)의 공격력에 맞게 감소하도록 함.

+++ Game/Source/Game/Weapon/Weapon.h
@@ -49,7 +49,7 @@ public:
 /*** 무기 스텟 설정 : Start ***/
 public:
 	UPROPERTY(EditAnywhere, Category = "StateMent")
-		int AttackPower; // 공격력
+		float AttackPower; // 공격력
 	UPROPERTY(EditAnywhere, Category = "StateMent")
 		float AttackSpeed; // 공격속도
 	UPROPERTY(EditAnywhere, Category = "StateMent")
